requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
stripe==7.12.0
python-dotenv==1.0.1
